+++ app/Http/Controllers/PublicController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Category;
 use App\Place;
 use App\Region;
 use Illuminate\Http\Request;
@@ -19,5 +20,13 @@ class PublicController extends Controller
     {
         
         return view('index')->with(['places' => $region->places()->where('is_approved', true)->get(), 'selectedRegion' => $region, 'is_regional' => true]);
+    }
+
+    public function indexRegionalCategories(Region $region, $category)
+    {
+        $category = Category::where('slug', $category)->first();
+        $places = $category->places()->where('is_approved', true)->where('places.region_id', $region->id)->get();
+        
+        return view('index')->with(['places' => $places, 'selectedRegion' => $region, 'is_regional' => true, 'category' => $category]);
     }
 }