+++ webpack.mix.js
@@ -11,5 +11,8 @@ const mix = require('laravel-mix');
  |
  */
 
-mix.js('resources/js/app.js', 'public/js')
-   .sass('resources/sass/app.scss', 'public/css');
+mix.js('resources/js/backend.js', 'public/js/backend.js')
+	.js('resources/js/frontend.js', 'public/js/frontend.js')
+	.extract(['vue', 'jquery', 'lodash', 'popper.js', 'axios'])
+	.sass('resources/sass/backend.scss', 'public/css/backend.css')
+	.sass('resources/sass/frontend.scss', 'public/css/frontend.css');