+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-Changelog is kept in english in an effort to make the whole projet open to more people.
+# 1.5.0 (2020-03-22)
+- Divided js and css resources in two parts (frontend and backend), and extracted vendor files.
+- Can filter region listing by categories.
 
 # 1.4.1 (2020-03-21)
 - Moved to a Github organization for publicly announced open sourcing of the project, so we cleaned up our readme, changelog and contributing guildelines.